Adding Install in its own section
After having a friend try to install this and getting stuck, I realize that there is no install section. Having it at the top makes running this a bit easier - otherwise, you have to search for install to learn the name of the repo. This should help onboarding new users, with minimal amount of space.