Add GitHub Enterprise setting as env var
Here, I add process.env.GITHUB_ENDPOINT as a possible environmental var
to use to change the rootURL. I change all instances of `rootURL` to
`githubEndpoint`, as that is the token used in [`gh-got`](https://github.com/sindresorhus/gh-got)
and it seemed like a good idea to just copy that token.

I was unsure of how to add a test for this.